base/base_trainer.py: Add a parameter for the no. of samples in the evaluation
base/base_trainer.py: Add a parameter for the no. of samples in the evaluation
trainer/trainer.py: Add a parameter for the no. of samples in the evaluation

test.py: Add a parameter for the no. of samples in the evaluation

config_baseline_SSD.json: Add a parameter for the no. of samples in the evaluation
config_learnt_SSD.json: Add a parameter for the no. of samples in the evaluation